Separate Game logic into corresponding files. Change COMMAND_KEY event checks from e.keyCode to e.watch (better support).
