+++ client/src/App.css
@@ -32,7 +32,7 @@
     align-items: center;
 }*/
 
-.ascii {
+/*.ascii {
     border: thin solid orange;
     box-sizing: border-box;
     height: 100%;
@@ -42,12 +42,12 @@
 .ascii-button {
     height: 50px;
     width: 300px;
-}
+}*/
 
 .screen {
     width: 100vw;
     height: 100%;
-    border: thin solid orange;
+    border: thick solid orange;
     box-sizing: border-box;
     display: flex;
     justify-content: center;
